fix(ingestion): use transaction client for feature flags
We were not reusing the PG client for the update of feature flag
overrides before which I believe is not desired behaviour.

We were investigating the issue of the UPDATE posthog_cohortpeople query
being the last query to have run in the transaction which was idle and
noticed that the feature flag overrides query is being run immediately
after, which seemed odd that we didn't see _that_ query as the last one.
It's possible that you get into a state where e.g. the pgpool is
saturated, we run the cohort query, then the feature flag query tries to
run, but there are no connections available, so it waits for one to be,
but that never happens.