chore(retries): reduce retries to 3, with 3 seconds initial delay
The way we are handling reitres at the moment isn't ideal. We end up
blocking a whole partition for a long time. Ideally we would push the
failed messages to a retry topic and have a delayed consumer on that
topic.

However, that's a larger change, and this is a quick fix to reduce the
impact of retries.