chore(plugins): load VMs on demand to reduce memory footprint
It's possible this is misguided and there are other better places to
tackle this. At any rate, we load all plugins no matter if there are
events that need to be processed by them, which is a little unnecessary.

I haven't really considered yet how this will affect e.g. plugin setup
and error handling in that respect.