refactor(s3-batch-exports): Swap to asyncio s3 client
Temporal activities are not heartbeating. I believe this is because the heartbeat
is spawning an async task, but since our code is not doing any await, we never
yield the main thread for it to run the async task.

By swapping to aioboto3 we now await on each upload part which should allow us to
heartbeat while the part is uploaded.