fix(s3-batch-exports): Call complete after uploading last part
Timing out doesn't raise an exception, so we are calling complete
before being done as the S3 upload context manager only looks for an
exception to be raised.

Instead, we now switch to manually calling complete after the last
part is done.

Not idea, but a bigger refactoring of this involves parallelizing s3
uploads and moving the context manager to the workflow, where it will
have visibility of the timeout exception.