Fix calendar range end
Use the end of the day (= 23:59:59) as the range end for the calendar
range picker.

The reasoning is that we probably don't want to adjust all queries
to treat just a day that is given as then end as the full day.
Rather this change of the date picker seems pretty straightforward.