Preventing mouse wheel from altering value in number fields
A user pointed out (in ZenDesk ticket 11999) that the value in the port number field, in the batch export config screen, can be changed unintentionally by scrolling with the mouse wheel.  This PR is to prevent that, and to prevent up and down arrow keys from altering the value since the user likely intends to scroll the page rather than increment the value in the field.