Wrap asgi application to ignore lifetime requests
ASGI servers send "lifetime" requests to django at startup and shutdown.

The ASGI spec allows these to fail, in which case the server ignores them, however
the way django fails them sends an exception to sentry which we don't want.

Overwrite the handler to return a 501 error without throwing an exception