fix(survey): filter survey responses count by earliest survey created in a team (#22011)
Resolves #21775

This PR addresses an issue with how we calculate response counts for a given survey in the surveys UI.
Instead of filtering out all events for a team by the event name of survey sent, we will look at only those events that are created after the team created their first survey. This should be a good indicator for when we should start to see survey responses for a given team.