feat: exception events don't update person properties
https://posthog.slack.com/archives/C0374DA782U/p1719318556646959?thread_ts=1719222351.040789&cid=C0374DA782U

Exception events will start going to a different pipeline, which might be out of sync with analytics pipeline, we don't want it to override properties to older values from here if it that pipeline were to lag. Exception events don't need to update person properties, analytics events do.

We'll still use `$set` to override the values for PoE for the exception event.