fix(cdp): Change StreamName -> StreamARN in Kinesis template
https://docs.aws.amazon.com/kinesis/latest/APIReference/API_PutRecord.html#API_PutRecord_RequestSyntax

The inputs ask for an ARN but the code populates the API's Name parameter. Best I can tell, this was mislabeled even in the first iteration of this template, and right now it fails in an obvious way in the function test logs.

You could use the StreamName, but Kinesis recommends the ARN. I think it's also nice to see the account in which the stream was created.

```
When invoking this API, you must use either the StreamARN or the StreamName parameter, or both. It is recommended that you use the StreamARN input parameter when you invoke this API.
```