Filter visits by language
This is doing some magic to filter by language. It's taking the browser language and then moving on to get just the language from the locale (i.e. pt from pt-BR). Because we want to be smart with the flags, however, we are collecting the country with the most views out of them. This should guarantee that people will see a British flag for en if most of their userbase is in the UK or an American flag if it's in America (or India, Nigeria, Australia, etc.)

There's still a small bug where filtering won't work if you simply click on the row, that needs to be updated.

See #26376