Add `CompareFilter` to Web Analytics
This filter isn't working yet, so the filter is hidden behind a FF. We've set the state such that the default behavior will be the same as it was previously.

A follow-up commit will update our `WebOverview` and `StatsTable` queries in the backend to use the values coming from the frontend.