feat(web-analytics): Support diff on Web Goals dashboard
We were not computing diff for the web goals section of our dashboard. This wasn't detected earlier because we don't have any actions on our local setup - we should change that.

Besides adding diff, we're also now properly formatting it in the UI, it was previously using the old formatting, therefore percentages were not accurate.