Fix: Update feature flag insights when key changes
This only updates the insights that the PostHog system creates when a feature flag creates.

If a user adds another insight to the feature flag dashboard, it's on them to keep it updated.

If the user changes the name of a system created insight, then we will not update it. We don't want to stomp on user changes.

Fixes #24938