feat: Allow Web Analytics to filter by cleaned path
We're now allowing us to filter by cleaned path, which makes our UI slightly more powerful. If you select a cleaned path and `doPathCleaning` is set, we can see results from a specific cleaned path. The UX isn't great because it looks kinda wonky, and I'm pretending it's a `EXACT` match (which kinda is?) and you can't get into that state unless you click on a path in the UI, but it does work, so I think this is a win. Eventually once we have `$pathname` a high-level filter (rather than just another filter) we'll be able to make this behave better.

Web Vitals is coming in the next commit