refactor: Improve event debug menu performance
When we had more than 2 events, our event debug menu was PAINFULLY slow. This is because we'd have several instances of `AnimatedCollapsible` rendering the `SimpleKeyValueList` even though they were usually collapsed. Let's remove the animation, it's not helping us on anything besides making my M3 Pro run at 3 FPS which is completely unreasonable. The height calculation is also kinda broken when devtools are open, this is fixed now.