fix: Properly handle empty `math_quantile`
Previously we'd throw an error of `Illegal type String` if we tried computing the percentile for an empty lsit of values. The problem would not happen for `avg` so I got interested. Turns out that we're not doing the same edge case handling for quantiles.

I've fixed it by handling it the same way, by forcing a cast to `float64` before computing the quantile. No more errors anywhere now! Nice!