feat: Collect host definitions from the `propdefs` repo
Web Analytics wants to, eventually, allow you to select which domain (host) you want to see your data for. People can already filter by it using the `host` field, but we don't have an easy way to display a dropdown of all available options.

We ask people to list their hosts to be able to access the Toolbar, but that could be 100% automated if we paid attention to the list of domains we're getting emails from.

Hence this commit: let's start parsing all of the hosts we see when parsing the definitions and create an entry in a brand new `hostdefinition` table. We aren't using the table yet, but we'll do so eventually.

I need to check with the infra team if we'll be able to keep up with the extra processing. It doesn't feel that heavy by looking at it, but I don't have much information on how close to the tipping of our service we usually are.