feat: Optimize layout for smaller laptop
In some smaller laptops the chart is under the fold which isn't ideal. I've moved a lot of the options under an "advanced options" dropdown. I've included the new goal lines but also some existing settings that don't need to be toggled often.

I've revamped the types as well because this is was extremely outdated. There's no more concept of left/right, nor anything to do with count.