feat: Introduce sortable Web Analytics tables
We can now sort our web analytics tables. I didn't use the existing sort
feature because it relies on tweaking the `query` which isn't something
we want here, we need to change the state inside the `webAnalyticsLogic`
and then expose that inside the tiles so that it can properly refresh.

Backend implementation is a little finicky but I believe we won't have
any problems. Tests cover almost all possible cases and they are passing
just fine.

I'm hiding this behind a FF but I intend to enable it 100% from the
get-go, this is just a way to easily revert in case UI breaks heavily.