feat: Add `HeatmapButton` cross-sell button
This will allow us to push people from Web Analytics to Heatmaps. Even though this isn't a widely used feature, we'll be pushing people to try the toolbar and it'll make sense once we start making Web Analytics available in the toolbar.

We're doing a big hack to decide which domain we want to redirect people to, but it'll become easier once we're filtering Web Analytics by a specific domain - coming soon, probably next week.