fix: Start looking at HogQL functions for geography detection
We were previously looking at plugins to detect if people had access to the world map or not. Let's now look at the hogql functions instead, because that's where they live.

Our basic check is by looking at the template that created the HogQL function, similarly to how we're looking at URLs previously. We have a failsafe logic to look at the name in case someone implemented this manually (unlikely but possible)