feat: Add `hit_count` to `hostdefinition` table
Turns out some customers - such as ourselves - might send events from
several different domains. This might be because of preview environments
or, in our case, something like our toolbar being used in someone else's
domain.

This is not ideal if we want to use this list to display as a dropdown
in Web Analytics. We need to find a way to figure out what the
"current" domains are.

To achieve that, we'll do two-fold:
- Look at `last_seen_at` to only display the ones that have been seen
  recently
- Add a new `hit_count` column so that we can detect non-ephemeral
  domains, ignoring preview environments
- We'll also ignore known `.vercel.app` and similar domains once we
  implement that dropdown