layout: Revamp Web Vitals layout
It was hard to make Web Vitals look right, and I could never find the right set of surface-colors to make it look good both on dark and light mode.

To solve that, let's simply make it look much more like the rest of the app. White highlights for our surfaces make it look so much better.