feat: Deprecate `host_definition`
This was an attempt to collect all domains so that we could use it on Web Analytics' UI. We've decided to go a different route because of the problems we'd have with having this inside `propdefs`. We don't usually drop any tables from our DB, so let's simply tag it as deprecated.

We've also implemented a new `@deprecated` decorator for python for this usecase, we're using it for a different already-deprecated class as well