feat: Add exchange rate table and dictionary
This will be used by Web Analytics to convert revenue between currencies. There's a `historical.csv` dictionary that we'll use to backfill everything since January 1st 2000 all the way to the end of 2024. Everything after 2024 will be generated via a Dagster job, it'll help us guarantee that this is working correctly.

Not every value is filled in, this is a sparse table, implying that you need to look at the closest date smaller than your current date if you wanna know the current rate. We'll have a hogQL function that will handle that. This is not only required for historical data, but also in case we're in a new date and we haven't gotten any rates from the new day yet.