feat: Add graceful-er shutdown handling to import_data_sync
Essentially, there are 3 ways to handle a worker shutting down:
1. Ignore it (not handle it, basically).
2. Raise a retryable exception.
3. Save some state, and raise a retryable exception.

Given that 1 is explicitly **not** handling it, and 2 and 3
culminate in the same exception, there is really only one way
to gracefully handle a worker shutting down: raising a
retryable exception.

Currently, we are doing 1 (nothing). This PR introduces a new
`ShutdownMonitor` that can be used by cooperating activities to decide
when to raise a retryable exception.

Work has been done to add such usage to the `import_data_sync`
activity.