feat: Add `hogql_convertCurrency` HogQL function
This function is meant to be used to convert from a specific currency to
another for a given date. This uses the newly-created
`exchange_rate_dict` to trigger the conversion. We don't have dates for
every single day yet, but they're coming soon, this should be enough to
test it out in prod!