feat: Add basic data exchange tables configuration
We're now letting customers select a specific data warehouse table to display as revenue. This is not very useful yet because Stripe is the only integration that actually makes sense as revenue, but it'll grow eventually.

We're not using that information in the backend yet either, it's coming soon. There's a small problem where we'll need to handle rows/events that don't contain a session ID, so that might be slightly tricky.