feat: Display revenue rows from our database
We're now querying the DB and looking for events in the DW tables. Main
concern right now is how do we limit it/sort by a specific timestamp to
get this to work/paginate properly? We might need to find a way to
figure out what the right "cursor" for a DW table is