fix: Use much improved Data Warehouse selector
This looks much nicer now, it's based on how this works for Product Analytics, and the behavior is definitely much nicer. Backend was kept basically the same, only change is that we altered the property from `name` to `tableName` to be both more descriptive AND match the existing `eventName` for events

We've changed how pydantic/the Team model represent our schema slightly, the behavior is much nicer now!