fix: Properly cast to `Decimal64`
A couple months ago we changed our `toDecimal` function to use `toDecimal64OrNull` but that brought us a problem where it only accepts strings as first argument, we can't do something as simple as `toDecimal(10, 1)` because it'd complain the first argument is a Float.

To fix that, let's go back to using `accurateCastOrNull`. The downside is that the way we decide on the scale is pretty weird, it looks like this `accurateCastOrNull(10, 'Decimal64(1)')` inside a string. We can easily add the extra string parameter by using `suffixArgs` (which only really exists because of `accurateCastOrNull` usage) but we couldn't easily add a value inside of it. That's solved now. It's full of restrictions (you can only pass constants, for example) but it works pretty well, see regenerated snapshots.

While I was at that, I changed `DATE` to `_toDate` to match our other functions, this was a mistake from the past.