feat: Introduce required `revenueCurrencyProperty`
It's hard to test this feature out if you can't say, for example, that
all of the events coming from a specific type of event have a set
currency that's different than the base currency. We could allow people
to write HogQL here, but this is a first step towards that.

I've updated both UI, server and tests in a single commit. I know this
shouldn't fit on the Data Warehouse PR, but it's the easiest way to get
this out and not have tons and tons of PRs open at the same time.