Add new top-level `flags` property to `/flags`
This ports the python implementation for `/decide?v=4` to Rust.

Add a `LegacyFlagsResponse` and a `ServiceResponse` enum so we can version the response.

By default, we use the logic for our `?v=2`, but if we're responding to a legacy request, we can derive `feature_flags` and `feature_flag_payloads` from `flags`, so we do that.