fix: Consider currency on revenue chart
We were not doing any currency conversion on the revenue chart, let's start doing that now

We had to add a new `math_property_revenue_currency` property to insights which is a bit too much, but I consider it to be ok.

I've also added that to the `column_optimizer` to make sure we consider that when thinking of materializing columns for our customers.