feat: Include revenue from data warehouse on revenue chart
On Web Analytics we're now including data warehouse events when computing revenue on the line chart (which will be a stacked chart if using events and data warehouse)

Because we've had to change the JSON architecture but I wanna force all fields to be set I'm resetting all configs in the DB. This is behind a FF that only our own account has access to, so I don't mind changing it. I'm resetting it at the DB level to guarantee all goes according to plan, though.