feat(batch-exports): Ensure only one ClickHouse query per Workflow
ClickHouse has multiple limits, like on resources and running queries.
We should ensure that a batch export doesn't create multiple
ClickHouse queries per run, and that it doesn't leave any ClickHouse
queries running after it is done.

This commit attempts to kill any running ClickHouse queries using the
Workflow ID as the query ID. This ID is unique to the batch export
run.

We attempt this kill in two moments:
* When exiting the `ClickHouseClient` context manager.
* When starting a new batch exports `Producer`.

The first case ensures we clean up after ourselves.

The second case ensures that if there was a problem during clean up,
we clean up on a retry.

However, there is one edge case that is not covered here: Activities
can keep on running after timing out as they are not informed of this
by the worker. These "zombie" activities will still have a running
query, and upon killing it will fail. However, they will then attempt
to kill the same query (as per the first case), which can lead to the
new attempt failing too, and potentially entering a killing loop.

This will be addressed in a future commit.