fix(temporal): Wait for worker PID in loop
`wait` will return immediately after a signal is handled by `trap`
with a status greater than 128. This means we will stop waiting on
worker PID, even if it has not finished yet.

Wait in a loop, checking if we have returned due to signal handling by
looking for status greater than 128. If we handled a signal, then keep waiting.