fix(external-data-job): Retry on incremental jobs when shutting down
Incremental sync activities are configured to not
retry (maximum_attempts=1) whereas full-refresh activities are
configured to retry 3 times. I think this should be the opposite: we
want full-refresh to keep on running and try to get it done once,
whereas incremental retries can fail and continue in a retry.

For this reason, I've switched retries: Incremental syncs now retry up
to 3 times, and full refreshes only have one attempt.

Moreover, I am only raising an error when shutting down if we are
running an incremental sync. Full refreshes will be allowed a chance
to finish, even when shutting down.