fix: Fix typecast to `Nullable(Date)` on warehouse revenue
Because these columns can be `DateTime`, HogQL calls `toDateTimeBestEffort64OrNull` on these columns. The problem is that we can't pass a `Nullable(Date)` to `dictGetOrDefault` because it expects it to be able to convert to `Uint8` which `Nullable(Date)` doesn't.

There are 2 approaches to this:
- update `convertCurrency` to always convert to a `0` `Date`
- update the call site to convert it to a non-nullable `Date` only when needed

I'm choosing to do the latter. It's more future work, but it means we need to be intentional about it which is much better IMO