refactor(snowflake-batch-export): Use a semaphore instead of timeout
If there is indeed a deadlock happening between multiple batch exports
connecting, the task in charge of monitoring the timeout will also be
blocked, thus we need a different approach.

Instead, use a semaphore that will prevent the blocking code from
running if anybody else is already running it. This limits concurrency
to only 1 batch export at a time can connect to Snowflake, thus no
possibility of deadlocks between 2 or more.