feat: Add basic data warehouse structure
This should let us test the new data warehouse structure for revenue
analytics in our test page (inside data management).

We'll not be adding these to web analytics because it's just too
tricky: getting access to that data *from* the frontend is tough, we'll
need new query runners to make this work well.

We can still relay to `InsightsQuery`s but it will need to be done in
the backend - like Web Vitals, for example.