feat: Add Revenue Analytics v0
This our first attempt at Revenue Analytics. This is just the initial setup, configuring all of the file structure and attempting at using a virtual `view` to display this inside PostHog. We arent adding any charts yet, this is coming soon.

We'll probably need to create cross-table views (for invoices and charges) but that's coming in a future PR, this is only the first PR to get an idea from the Data Warehouse team whether this makes any sense or not