chore: Update manifest.toml to run data warehouse tests locally
We're storing our data warehouse data inside `objectstorage` locally but that's not properly resolved by default because we didn't add it to `/etc/hosts`. There are 2 solutions to it:

- run tests inside docker
- expose that outside docker via `/etc/hosts`

We're doing the latter :)

I've also fixed the script slightly. We were previously only appending that to the hosts file if running on an interactive shell, but that's not right, we should always run it