fix(batch-exports): Support all event fields in custom query
When passing a custom hogql query, we were losing some node
information while parsing as the dialect was set to hogql on the
initial parse. Now, we set it to ClickHouse to avoid losing node
precision. This allows us to support fields from lazy
joins (person_id) if they are present in events.