refactor: Improve InsightLabel
The label now wraps in a much better way by forcing the `MathTag` to be all in the same line. It'll keep in the same line if there's space - we've given more space to it (from 120 to 200) - but wrap as a whole to the following line if there isn't space to it + the actual label in the first line.