fix: Reintroduce managed views to the database
We had to remove them because of a bug, see https://github.com/PostHog/posthog/pull/31271

We're now adding them back. The whole bug is because we had both `stripe` as a means to store data warehouse tables but also for views - because of our dot-notation approach.

Because we call `database.add_views` after `database.add_warehouse_tables` we were basically overwriting the previous `stripe` entry.

We're now much smarter and we actually *merge* the `TableGroup`s together.

Once that was fixed we also had to update the views handling code to not include something like `stripe` to the list of views. This is already properly handled for data warehouse tables.

More context https://posthog.slack.com/archives/C019RAX2XBN/p1744762144664099