feat: Improve month handling for revenue
We're now properly using the dates passed in as `dateRange` for our revenue queries.

Frontend was also updated to present a more sensible set of defaults for our timeframe picker - removed the interval because we wanna control it ourselves.

We still need to avoid querying *some* data based on the date selection AND properly adding a toggle between `month` and `all` for the top customers feature - toggle between a table and a chart view.