refactor: Use fixed fallback for revenue query data
When running a query with the `all` timeframe, we usually run a query to figure out when the earliest ingested event happened. That doesn't make any sense for data warehouse tables, so let's change this slightly.

For simplicity, let's simply assume January 1st 2015 for now. Some people might ingest historic data, but we can cross that bridge once and if we get there