feat: Filter sources/events in the UI
We're now letting customers filter which sources AND events they wanna see data about in the revenue UI. We've both added that filter to the UI and updated the backend to respect that feature.

Backend is looking kinda weird now, but we can clear this up soon - splitting `revenue_analytics/backend/models.py` into a couple files.