feat: Use new `TeamRevenueAnalyticsConfig`
We've now replaced the old JSON-based config with this new model, this should make our code much easier to read and reason about. API is also much simplified and improved, much easier to get things right too.

There's a migration here that will backfill all of the existing config, this should run quickly after we merge the previous PR and create this table. There's a small chance people will update their settings in the meantime, but it's all right if they dont match, this is easily fixable by customers if they try again.