refactor(dwh): New config abstraction for source configuration
This abstraction can be seen as an extension of Python's builtin
dataclasses. By decorating classes with @config we automatically gain
methods to process configurations coming from user requests and stored
in our database. Moreover, all the usual dataclass stuff still works.

This will be used to replace a lot of the manual configuration
handling we have duplicated throughout our codebase.