Add project API token resets to activity log
Log changes to the project API token in the activity log.

Project API tokens are not secrets. They are public to the website when used by customers.

They are especially not secret to the team because they're clearly visible on the project settings.

And if someone were to change the key, I'd definitely want to see that in the activity log as it could break so much.

That's what this commit does.