fix: Allow `WHERE {column} IS NOT NULL` in index creation
This check is trying to protect against adding a column with a non-null constraint to a table OR adding a non-null constraint to an existing column.

However, the only way such a thing can occur is if we're altering a table. So it seems changing the condition to look for the existence of "ALTER TABLE" should be safe and more robust.