Add a custom permission for project secret api token
We still have to update the other permissions to ignore requests authenticated by our new authentication method
because permissions are additive. So to add this new custom permission, we're still calling all the other permissions.

I could factor out these two endpoints to their own ViewSet, but that's a lot of churn and I'm already porting local_evaluation to Rust.