Remove product_description from demo data
We can still keep some vestigial support for `Project.product_description` for earlier users of Max, but we should drop it some time after launch, and definitely shouldn't be including in the demo data anymore. Instead, the demo data should have either a core memory pre-set, or nothing. Opting for nothing here, so that we experience the onboarding actually.