fix: Fix InfiniteList logic when using `showNumericalPropsOnly`
Because of the way we do filtering + rendering in our UI we had an inconsistent loading state for lists using `showNumericalPropsOnly`. The backend would return a very high count of rows - which we always render in the frontend - and we would not tag them as "void" when we filter them out, causing us some troubles - people thought we were still loading when we weren't anymore.

Let's fix this by simply querying by numerical props only when the prop is set.

A follow-up PR will introduce a FF to remove the filtering in the frontend, it might be unnecessary - need some testing in production.