feat(hogql): Support modifying csv allow double quotes settings
The format_csv_allow_double_quotes ClickHouse setting is defaulted to
0. But, as CSV is a very shaky format, this can cause lots of trouble
when CSVs are not formatted in certain ways, particularly when JSON
values are nested in CSVs.

We should ultimately allow users to customize how do their CSV files
look like, but it seemed faster for now to allow configuring this on a
team by team basis.

So, I've added a new HogQLModifier that when set will update the
aforementioned setting.

The plan is to then set this modifier for teams who request it.