Fix the test
Cohort queries no longer looks at the `person_properties` column on the `events` table because those values get out of date. Instead, it looks directly at `person.properties`. So if we materialize a column, we need to do it on the `person` table, not `events`. After doing so, the query leverages the materialized column.