fix(zendesk): Add prefix to configuration values update
Zendesk source requires a `zendesk_*` prefix on all configuration
values. For some reason, our frontend is sending values without this
prefix, and on `update` we do not add the prefix like we do in
`create`.

This solves the problem in a bit of a rushed way, ideally we should
refactor the API to use a proper configuration class, but I am letting
that for later.