feat: Implement new Rev ANalytics insights query runner
Rather than using the default insights, let's instead use a custom one where we can customize it in the backend. It will expose the same interface as the insights currently do to simplify using/building new frontend components. For now we're just rendering the JSON, but we'll soon reimplement and display the same chart we had previously.

Follow-up commit will also let us choose between the existing insight and a per-product split instead.