feat(spmc): Split queries on OOM
ClickHouse memory limits can cause batch exports to fail. As a
potential solution, we can split a query that fails due to memory
errors into two queries with hopefully smaller footprints. Eventually,
these two queries could be run concurrently, as the query limit is per
query, not global, but that would have consequences with ordering so
for now it's just sequential.

This feature is only available for queries that have interval start
defined, so for every query except "from the beginning of time"
backfills. This is left as a TODO for later. Most queries are not
backfills anyways, so I am hoping this solution is enough.