feat(batch-exports): Configure virtual style addressing in S3
Some non-AWS but S3-compatible destinations require virtual style host
addressing. This is configured in the boto3 client, but we currently
have no way to set this configuration. So, this adds a new setting to
both the UI and the backend to allow configuring this specific
setting.

In the future, we could expand this to support the full breadth of
configuration options, but for the time being we only need this, so
I've kept the changes focused.