feat: Allow Revenue Analytics data to be filterable
This is a crazy commit, but it's impossible to break it into smaller bits, the code won't work unless we change all of this at once. There's a big change from using `charge`s as the default view to using `invoice_item`s instead (`invoice.lines` for those who know how Stripe works) which is much more reliable and gives us much more information.

This basically allows us to filter the data we're seeing in the Product Analytics dashboard. Currently you can only filter by product and amount, but more filters are coming soon now that we have the tech for it.