refactor(google-ads): Hardcode schemas to allow metrics and segments
Google ads schemas can be heavily customizable with metrics and
segments. This commit makes the schemas match the reports that google
ads exports to BigQuery by default.

Ultimately, we should allow users to customize their reports, but for
the time being this provides a baseline experience in line with what
is offered by google ads natively.