feat: Filter Rev Analytics by source using generic filter
Rather than having a custom filter to choose which events/sources we wanna use for displaying our data, let's instead offload it to the same filters we're already using for the rest of the ui.

This will also solve a lot of bugs we had where we loaded this page before we'd finished loading sources from the backend.