refactor(spmc): Reduce calls to record batch slice
Slicing a record batch is a zero copy operation, so no data is moved
around. However, we still pay some cycles for each call: In my
testing, execution time grew linearly with the size of a record batch.

This means we can optimize the `slice_record_batch` function if we
reduce the number of calls to `pyarrow.RecordBatch.slice`. The current
loop it's quite hot as it is taking 1 record out of the slice at a
time. Instead, we can take multiple rows at a time, by estimating how
many bytes we are still missing from the target.