Fix the order of data points displayed in the chart tooltip
The data points were being rendered in the reverse order and this was causing issues when there were undefined data points for a series.

By removing the sorting, we default to the sorting as returned by the server, which is already in the expected order.

For more information about the issue check https://github.com/PostHog/posthog/issues/33209