feat: Initial work for virtual `revenue` fields for `persons`
We want people to have access to revenue data on the `persons` table by doing something like `persons.$virt_revenue`. This PRs achieves this by creating a new lazy table called `raw_revenue_analytics` that we can then join on. We're currently hardcoding the join on the Stripe ID which looks like `cus_xxxx`. This will only work if people are aliasing their PostHog persons to the ID they have in Stripe - which is unlikely. We'll make this configurable in the next commit.