refactor: New transformer interface for batch exports
This commit changes the `StreamTransformer` interface introduced for
batch exports.

First of all, I am now using a protocol instead of subclassing. The
name of the new protocol is `Transformer`. The reality is that each
file format is it's own thing, and protocols give us more flexibility.

The `Transformer` protocol is really simple: Just requires a method
that iterates over `Chunk`s.

Introduced a new tuple `Chunk` which indicates the bytes we are
streaming and whether we are at EOF.

Brotli now does multiprocessing too, but compression happens in the
main process.

I think files may have been not finishing at the correct place if the
record batches were too large, so this corrects that too.