fix: Make $virt_revenue fields work
The basic problem we had is that we were exposing the revenue analytics table inside `persons` but that doesnt make sense because we need to connect revenue with a specific `distinct_id` instead, therefore it should be part of the `distinct_id` table, which we've done now.

There's a small problem with our approach, currently: we can properly display this field as a table, but it fails when trying to filter because we have a WINDOW function on a field that goes to the `where` which isnt supported - weird, but ok.

We need to find a better way to expose this field in the parent `persons` table such that we only have a *single* value in it without using any kind of aggregation function that would then fail for two-fold: not using group by, not being able to use window on WHERE.

Subquery comes to mind, but it's tricky to get them right.