feat: Start calculating MRR/Revenue in the `GrossRevenue` query runner
We're now calculating MRR alongside revenue in our existing query runner. We're doing most of the calculations in Python-land because it's so much easier than it is to compute it via Clickhouse. It's also pretty fast because we're only iterating over the grouped results.

Follow-up commit will update the UI.

The name will need to change, I know