feat: Reimplement `revenue_analytics` table using `person_id`
Rather than implementing this table at the `pdi` level with the `distinct_id`, let's instead do it at the `persons` level. This sounded tricky at first, but we can easily achieve this by exposing the `person_id` from the `customer` view in Revenue Analytics. On the revenue events, also expose that and simply join on that field, group by it, sum the values and boom, easy!

Thanks to @mariusandra for the idea, worked like a charm