feat: Add external logger for user-facing logging
Currently, it's not clear which logs go to external consumers (i.e.
users) and which ones are meant only for internal use.

This introduces a logger, named "EXTERNAL". Only logs from this logger
will be put into the external logging pipeline. All other loggers will
be passed along without being placed in the queue.

In order to not break other setups, there is a new setting that
enables this, otherwise the previous behavior is maintained (whatever
that is, it's not clear to me how things work right now).