refactor: Configure logging on worker start-up
There shouldn't be any need to guess whether logging is configured and
juggle that around through the lifetime of the worker. Just configure
it once when we start and put it out of our way.

This is only implemented for the async logging pipeline, and gated
behind the `TEMPORAL_USE_ASYNC_EXTERNAL_LOGGER` variable. More work is
required to support also the sync version (potentially, combining both).