fix: List proper `invoice_id`
Because HogQL is shadowing the `id` field, we were actually rendering the `invoice_item_id` field here rather than the `invoice_id`. We werent using this for any join, so that's not a big deal, but it's good idea to fix this in case people build on top of our views.